+++ pkgs/ffigen/example/objective_c/play_audio.dart
@@ -6,15 +6,12 @@ import 'dart:ffi';
 
 import 'package:objective_c/objective_c.dart';
 
-// TODO(https://github.com/dart-lang/native/issues/1068): Remove this.
-import '../../../objective_c/test/setup.dart' as objc_setup;
 import 'avf_audio_bindings.dart';
 
 const _dylibPath =
     '/System/Library/Frameworks/AVFAudio.framework/Versions/Current/AVFAudio';
 
 void main(List<String> args) async {
-  objc_setup.main([]);
   DynamicLibrary.open(_dylibPath);
   for (final file in args) {
     final fileStr = NSString(file);