+++ pkgs/ffigen/test/native_objc_test/native_objc_test.dart
@@ -17,16 +17,6 @@ import 'util.dart';
 void main() {
   group('native_objc_test', () {
     setUpAll(() {
-      // TODO(https://github.com/dart-lang/native/issues/1068): Remove this.
-      DynamicLibrary.open(
-        path.join(
-          packagePathForTests,
-          '..',
-          'objective_c',
-          'test',
-          'objective_c.dylib',
-        ),
-      );
       final dylib = File(
         path.join(
           packagePathForTests,